UF-10134: Fix sonar issue

+++ src/test/java/se/sundsvall/citizenchanges/scheduler/FileHandlerTest.java
@@ -33,7 +33,7 @@ class FileHandlerTest {
 	private FileHandler fileHandler;
 
 	@ParameterizedTest
-	@ValueSource(strings = { "Nej", "Ja, deltid" })
+	@ValueSource(strings = {"Nej", "Ja, deltid"})
 	void getISTPlacement_nej_ja_deltid(final String daycarePlacement) {
 		// Arrange
 		final var parsedRows = new ArrayList<ParsedRow>();
@@ -71,6 +71,7 @@ class FileHandlerTest {
 		assertThat(result.getDaycarePlacement()).isEqualTo(daycarePlacement);
 	}
 
+
 	@Test
 	void getISTPlacementWithDayCare_Heltid() {
 
@@ -101,33 +102,6 @@ class FileHandlerTest {
 	}
 
 	@Test
-	void getISTPlacementWithNoMatch() {
-		// Arrange
-		final var parsedRows = new ArrayList<ParsedRow>();
-		parsedRows.add(ParsedRow.builder()
-			.withPersonId("12")
-			.withUnit("someUnit")
-			.withPlacementStart(LocalDate.now().minusMonths(1).toString())
-			.withChangeStart(LocalDate.now().toString())
-			.withPlacementEnd(LocalDate.now().plusMonths(1).toString())
-			.withTaxCategory("Fritidshem heltid")
-			.withGuardian1("someApplicantIdentifier\nTest Testorsson")
-			.withGuardian2("someApplicantIdentifier2\nTest Testorsson")
-			.build());
-
-		final var item = buildOepErrandItem("Ja, heltid", "someApplicantIdentifier", "12");
-
-		// Mock
-		when(list.stream()).thenReturn(parsedRows.stream());
-
-		// Act
-		final var result = fileHandler.getISTPlacement(item);
-
-		// Assert
-		assertThat(result).isNull();
-	}
-
-	@Test
 	void parse() throws IOException {
 		// Arrange
 		final var lastRunPath = Path.of("lastRun.xls");
@@ -152,16 +126,16 @@ class FileHandlerTest {
 	void cleanup() throws IOException {
 		final var lastRunPath = Path.of("lastRun.xls");
 		final var mockFilePath = "src/test/resources/mockfiles/mockfile.xls";
-		if (new File(mockFilePath).createNewFile()) {
-			if (Files.exists(lastRunPath)) {
-				try (final var inputStream = new FileInputStream(lastRunPath.toAbsolutePath().toString())) {
-					Files.copy(inputStream, Path.of(mockFilePath), StandardCopyOption.REPLACE_EXISTING);
-				} catch (final IOException e) {
-					throw new RuntimeException(e);
-				}
-
-				Files.delete(lastRunPath);
+		if (new File(mockFilePath).createNewFile() && Files.exists(lastRunPath)) {
+			try (final var inputStream = new FileInputStream(lastRunPath.toAbsolutePath().toString())) {
+				Files.copy(inputStream, Path.of(mockFilePath), StandardCopyOption.REPLACE_EXISTING);
+			} catch (final IOException e) {
+				throw new RuntimeException(e);
 			}
+
+			Files.delete(lastRunPath);
 		}
+
 	}
+
 }
